pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "data-pipeline"
version = "0.1.0"
description = "A pipeline for NYC taxi data: import, filter, and predict tip amounts with Airflow integration."
requires-python = ">=3.8"

dependencies = [
    # Data processing
    "pandas>=2.1.3",
    "numpy>=1.26.0",
    "pyarrow>=20.0.0",

    # Database
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",

    # Machine learning
    "scikit-learn>=1.3.0",
    "joblib>=1.3.0",

    # Environment management
    "python-dotenv>=1.0.0",

    # Workflow orchestration (Airflow + Postgres support)
    "apache-airflow[postgres]>=2.10.0"
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.isort]
profile = "black"
